+++ src/FluentMediator.Microsoft.Extensions.DependencyInjection/FluentMediatorExtensions.cs
@@ -17,5 +17,35 @@ namespace FluentMediator
 
             return services;
         }
+
+        public static IServiceCollection AddSingletonFluentMediator(
+            this IServiceCollection services,
+            Action<IPipelinesBuilder> setupAction)
+        {
+            var pipelinesBuilder = new PipelinesBuilder();
+            setupAction(pipelinesBuilder);
+            var pipelines = pipelinesBuilder.Build();
+
+            services.AddSingleton(c => pipelines);
+            services.AddSingleton<GetService>(c => c.GetService);
+            services.AddSingleton<IMediator, Mediator>();
+
+            return services;;
+        }
+
+        public static IServiceCollection AddScopedFluentMediator(
+            this IServiceCollection services,
+            Action<IPipelinesBuilder> setupAction)
+        {
+            var pipelinesBuilder = new PipelinesBuilder();
+            setupAction(pipelinesBuilder);
+            var pipelines = pipelinesBuilder.Build();
+
+            services.AddScoped(c => pipelines);
+            services.AddScoped<GetService>(c => c.GetService);
+            services.AddScoped<IMediator, Mediator>();
+
+            return services;
+        }
     }
 }