+++ test/UnitTests/PingPong/PingHandler.cs
@@ -5,22 +5,22 @@ namespace UnitTests.PingPong
 {
     public class PingHandler : IPingHandler
     {
-        public PingResponse MyLongMethod(PingRequest request)
+        public PingResponse MyCustomBarMethod(PingRequest request)
         {
             return new PingResponse("Pong");
         }
 
-        public PingResponse MyMethod(PingRequest request)
+        public PingResponse MyCustomFooMethod(PingRequest request)
         {
             return new PingResponse("Pong");
         }
 
-        public Task<PingResponse> MyMethodAsync(PingRequest request)
+        public Task<PingResponse> MyCustomFooBarAsync(PingRequest request)
         {
             return Task.FromResult(new PingResponse("Pong"));
         }
 
-        public Task<PingResponse> MyMethodAsync(PingRequest request, CancellationToken cancelationToken)
+        public Task<PingResponse> MyCancellableForAsync(PingRequest request, CancellationToken cancelationToken)
         {
             return Task.FromResult(new PingResponse("Pong"));
         }